chore: dto 정적 팩토리 메소드 적용
